+++ CHANGELOG.md
@@ -1,4 +1,12 @@
 
+### Version: 0.1.3 
+#### Date: Feb-14-2020 
+
+- ContentstackHelper update with using statement
+- Update to Class implementation with 'partial class'
+- Added const string 'ContentType' to specify content-type of class 
+- Support to set reference field class in MultiCT for single reference 
+
 ### Version: 0.1.0 
 #### Date: Jan-10-2020